(API/APP) Alta de reseñas (falta validar y verificar)
Actualizada la base de datos.

+++ api/TiendaMovil/Models/Reseña.cs
@@ -15,7 +15,7 @@ namespace TiendaMovil.Models
         public int UsuarioId { get; set; }
         [ForeignKey("PublicacionId")]
         public int PublicacionId { get; set; }
-        public int Puntaje { get; set; }
+        public float Puntaje { get; set; }
         public string Encabezado { get; set; }
         public string Contenido { get; set; }
         public int Estado { get; set; }
